+++ History.md
@@ -1,6 +1,7 @@
 [3.5.0-pre / 2015-xx-xx](https://github.com/jakubpawlowicz/clean-css/compare/3.4...HEAD)
 ==================
 
+* Adds more detailed error & warning messages on top of the new tokenizer.
 * Requires Node.js 4.0+ to run.
 * Replaces the old tokenizer with a new one which doesn't use any escaping.
 * Replaces the old `@import` inlining with one on top of the new tokenizer.